pyproject.toml
[tool.poetry]
name = "polecat"
version = "0.0.10"
description = ""
authors = ["Luke Hodkinson <furious.luke@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
boto3 = "~1.9.125"
cached-property = "~1.5.1"
click = "~7.0"
coolname = "~1.1.0"
factory-boy = "~2.11.1"
graphql-core-next = "~1.0.2"
graphql-server-core = {git = "https://github.com/norman-thomas/graphql-server-core", branch = "master"}
halo = "~0.0.29"
polecat-feedback = "*"
psycopg2-binary = "~2.8.4"
pyjwt = "~1.7.1"
python-levenshtein = "~0.12.0"
pytimeparse = "~1.1.8"
pytz = "~2019.1"
sanic = "~19.3.1"
sanic-cors = "~0.9.9.post3"
termcolor = "~1.1.0"
ujson = "~1.35"

[tool.poetry.dev-dependencies]
pytest = "~4.3.1"
pytest-asyncio = "~0.10.0"
pytest-cov = "~2.8.1"
pytest-mock = "~3.1.1"
pytest-polecat = "*"
